gfs2_edit: Ability to enter "journalX" in block number.
With gfs2_edit, you can position to the "block #" field and
press <enter>, then enter a block number to jump to.
You may also enter a keyword, like "root" to jump to the
root directory, "rindex" to jump to the rindex, etc.
Most keywords worked, but you could not enter "journal0"
to jump to the first journal until this fix made it possible.