gfs2_edit: was parsing out gfs1 log descriptors improperly
GFS2 log descriptors have 8 bytes per entry which represents
a block number.  GFS1 log descriptors have a small structure
that is 16-bytes, the first 8 bytes of which is the block.
So gfs2_edit was mistaking the extra 0x00000000 for a GFS2
end-of-block marker when it shouldn't have.  This fixes it.