cman: add option to init script to prevent joining the fence domain
New sysconfig variable FENCE_JOIN will control whether or not the
node joins the fence domain. This variable can be set to either
value "yes" or "no". When FENCE_JOIN is set to "no", the init script
will not attempt to join the fence domain. Any other value is
equivalant to "yes", in which case the init script will attempt to
join the fence domain. (BZ #455598)