config: fix objdb2xml filtering
ais/corosync has a set of extra information on the toplevel
tree that includes <service service_id="" handle=""
that contains values for service_id and handle that cannot
be converted into XML data.

We were filtering for <service to avoid importing non-compliant data
but that was catching also /cluster/rm/service making it impossible
to use rgmanager.

Change the filter to match explicitly for service_id and handle.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>