libdlm: handle truncated device names
When lockspace names are over 15 characters long, they result in a
device name that's over 19 characters long, e.g. dlm_0123456789ABCDEF.
Sysfs truncates device names at 19 characters, so the device name for
this lockspace is /sys/class/misc/dlm_0123456789ABCDE, which udev also
uses to create the device node, /dev/misc/dlm_0123456789ABCDE.

So, when libdlm waits for udev to create the device node, it needs to look
for this truncated name.  It then creates and removes symlinks with the full
lockspace name.

Joel Becker identified the problem and came up with this solution.

Signed-off-by: David Teigland <teigland@redhat.com>