cman: add cman_tool -A to disable load of openais services
I have moved the loading of openais services to the control of
the openaisserviceenable configuration module (as it knows better
which services are available). Loading of this module, and
consequently the rest of the openais services is now controlled
by the -A switch.

openais services are enabled by default. -A disables them.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>