ccs: deal with xml file format special case
when talking about specific xpath queries, writing:
<clusternode name="node4" votes="1" nodeid="4"><fence>..

is not the same as writing:
<clusternode name="node4" votes="1" nodeid="4">
 <fence>

fix memory overflow in both ccsd and xmlconfig. It is impossible to
exploit this overflow for anything useful since it's used at the very
beginning of the  startup process when literally nothing is running
and it causes the XML parser to crash.

ccsd did never show this issue because it was using a pre-allocated
buffer and it was always big enough to hold the data (even when we were
writing more than our calculated size).

xml on the other side, always need to allocate.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>