GFS2: Make gfs2_fsck accept UNLINKED metadata blocks
bz 460327

Originally, GFS2 did not use a block type of 2 in the bitmaps,
so it was considered invalid.  However, GFS2 now uses that
block type to indicate unlinked metadata blocks.  This allows
for cases where an inode is unlinked on one node while still
open on another node.  This fix changes gfs2_fsck so that it
ignores these blocks (eventually the file system will reclaim
them) rather than reporting them as errors.