cman: honour the dirty flag on a node we haven't seen before
The dity-node code used to check if a node had been down before
honouring the dirty flag, this was to prevent nodes already in the cluster
from kicking each other out at a transition.

This had the problem that it could not detect if a new node joined that already had state (eg a new cluster started up in a split network).

So, now we also check the 'first_trans' flag in the transition message so
that we know when the node has newly joined a multi-node cluster.

For more information see bz#460909

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>