dlm/fence/gfs: fix daemon spinning 100% due to memory corruption
This is more a workaround than a real fix.

When building with -O0, arg list to pthread_create is somehow corrupted
(I suspect a gcc bug here as the problem doesn't show with any other -O
levels), and data passed down to process_query are invalid.

Stop passing arguments via pthread_create.

Add simple sanity check and fallback in process_queries on accept() call
that was the main cause of the spin.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>