gfs-kernel and mount.gfs2: GFS ignore the noatime and nodiratime mount options
Since the vfs moved the MS_NOATIME, MS_NODIRATIME flags
from the superblock to the vfsmount structure (MNT_NOATIME,
MNT_NODIRATIME, which are not accessible to gfs), gfs no
longer knows when to enable/disable atime updates and the
atime_quantum stuff is broken in the sense that it doesn't
respect the noatime and nodiratime flags.

This patch attempts to fix this by creating a gfs-specific
mount option gfs_noatime and having the mount.gfs helper
pass it along when the user specifies noatime or nodiratime
in the command line. It's not the ideal way to fix it, and
it is a bit ugly.