[fence] Operation 'list' for APC fence agent
Operation 'list' return list of available outlets (with aliases) or name
of virtual machines. This patch includes changes for fencing library and
changes for APC fence agent. There is third argument for fence_action()
that is None by default so we don't have to change other fence agents that
do not have support for this operation, yet. In some cases we can overload
get_power_status() as we just need to cache result in other cases it will
be completely separate function (eg. 'xm list').

@note: Current implementation for APC will list just outlets on one device.
If we have more than one switches (MasterSwitch) we will see outlets just
on this device.