ccs: add ccs_read_logging
this function is a simple wrapper to gather all logging configuration
from one central place since this code is replicated already N times
across the tree.

the function should always be invoked with the selected default values
and it will apply overrides from the config (if available and if
possible).

you are responsible for freeing char **file when changed via
cluster.conf.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>