[fence] WTI should not power on/off plug if it is unable to get status
Fix #468904. On some WTI devices plugs are numbered as 1,2,3 and on the others as
A1,A2,...B1,.... Both types accept numbers (A1 = 1, B1 = [number of last A] + 1).
Power on/off works with numbers but if we want to parse status of plug then
we have a problem. This patch is a general solution (fencing library) because
it tests value of get_status() which have to be on/off otherwise we exit with
new error code.