fence: Fix operation 'list' and 'monitor' for LDOM and ePowerSwitch
This two fence agents had same method for getting power status
and getting outlets. In method was simple checking, which
operation should be processed, but main fencing library doesn't
set this option to 'status' in case of powering outlet on or off.

This led to bad behaviour in case of on, off and reboot operation.
Operations status, list and monitor wasn't affected.