fence: IPMI over lan timeout adjusted and configurable
This patch adjust timeout to default value 10s which
should be enough for most today IPMI implementations.
It also removes retries, because this job is done
by fenced.

Because some devices still need longer timeouts,
timeout is adjustable by parameter -t (or timeout for stdin
and XML configuration).

This should fix BZ: 401481, 276541 and 452894