Fix many bugs with gfs2_convert.
bz 471618 - GFS2: gfs2_convert is broken

Functional changes to the code are as follows:

1. Original problem is fixed

The original problem reported in the bug was caused because
there is a fundamental shift of indirect block pointers between
GFS1 and GFS2.  For indirect blocks, GFS1 added 64 bytes of
reserved space after the gfs meta header.  GFS2 did not.  That
meant GFS1 could hold fewer pointers than GFS2 on the same block.
The meant code had to be added to gfs2_convert to shuffle all
the pointers around to their corresponding gfs2 locations.
This was not a simple calculation because of sparse files.  For
example, with a 1K block size, if you do:

dd if=/mnt/gfs/big of=/tmp/tocompare skip=496572346368 bs=1024 count=1

the resulting metadata paths will look vastly different for the
single block of data:

height    0     1     2     3     4     5
GFS1:  0x16  0x46  0x70  0x11  0x5e  0x4a
GFS2:  0x10  0x21  0x78  0x05  0x14  0x78

It is relatively easy to calculate the new metapath, but you
can't just shuffle a pointer from its old location to its new
location because the destination slot might be used for a
different source slot.  In the example above, we couldn't just
move the pointer at height 3 from offset 0x11 to another block
at offset 0x05 because there might already be a pointer at 0x05
being used for something else.

To complicate matters, we could not just assign new blocks and
copy the data because we should be able to run on a "full" file
system.  So I had to write a complex new function called
adjust_indirect_blocks along with several support functions.
Someone else might be able to write a function to get the job
done simpler, but at least this one works for both sparse and
fully-packed metadata trees of several heights.  It basically
formulates in memory a metadata tree for all the data blocks,
clears out all the metadata buffers, and lays them all back down
again according to the new layout.

2. Improved error checking

I added some badly needed error checking.  That's because I was
running into problems and wanted to eliminate the possibility
that errors were occurring, but not being reported.

3. Improved progress reporting

With a large number of resource groups, I saw long periods of
time where the program would just sit for five minutes.  I was
afraid that the tool had hung, so I kept breaking in with gdb
to check on the tool's progress.  But if I'm that concerned
about what it is doing when it's quiet, so too will be the
customers who are anxiously trying to watch its progress.  So I
added periodic reports of progress through the RGs when converting
inodes, and also an occasional "." when the RGs themselves are
being analyzed and/or manipulated.

Same goes for writing out the new gfs2 journals.  The journals in
gfs2 are quite different and we need to write them out entirely,
all 128MB.  That takes a fair amount of time going through the
functions in buf.c.  So I added a new message when each journal is
written, so that the code doesn't appear stuck for a long period
of time.  This could still use some improvement, but it's better.

I also modified the progress messages to make them more clear.
For example, the message "Removing obsolete gfs1 structures"
sounded to me like it could cause alarm or panic to a customer.
So I rephased it to "Removing obsolete GFS1 file system
structures".

4. Journal size was not being preserved

In testing, I discovered that if the source file system had 32MB
journals, the reformatted GFS2 file system had 128MB.  I decided
that's not good and decided to fix it.  That is especially
important if the file system is "full" to begin with because we
can't add more blocks.

5. Fixed a segfault dealing with buffer management.