[FENCE] Support for 'metadata' option for fencing agents / default values
Fencing library now understand new 'metadata' action which shows
available commands (for use on stdin). Output should be compatibile
with stonith agents. Currently it is just partial (parameters only).
We will extend it according to what we really needs. It should be
possible to use these metadata for building UI. Device options now
contains 'short description' and earlier option 'order' was added so we
can set the correct orded (it will be same as in --help).

Default values are now stored in fencing library as they are used for
metadata generating. Fence agent can set their default settings to all_opt.

There is a bit of redundancy ('required') because we will check it later.
But I was not able to describe requirements without using not-naive schema.