cman: add cman3 services
These services are corosync plugins and libraries that implement the 'old'
cman functions in a more standard manner. The quorum plugin is a generic
quorum provider that uses the CMAN (openVMS) algorithm to provide quorum and
adds the hooks for a quorum device. With a compilation flag it can be made
wire-protocol compatible with the 'old' cman plugin.

The new cman service provides everything else that iold cman providef that's
not related to quorum and is intended as a backwards-compatibility layer.

The libcman part is not yet complete so I'm not going to replace the existing
'daemon' directory here just yet, also I don't think it's ready to be built as
standard with the rest of the tree. This is more of a statement of intent
than working code ;-)

Signed-Off-By: Chrissie Caulfield <ccaulfie@redhat.com>