Medium (LF 2147): IPaddr2: behave if the interface is down
It is not exactly clear why should findif exit with error in case
if cannot find a route which may match the interface. At any
rate, IPaddr behaves different in such situations, i.e. it relies
on ifconfig. IPaddr2 now matches that behaviour. BTW, it is a
mess, badly needs a rewrite.