High: RA: introduce OCF_FUNCTIONS_DIR, allow it to be overridden (LF2239)
This introduces an OCF_FUNCTIONS_DIR variable, pointing to the
directory where the scripts find .ocf-shellfuncs and friends.
It defaults to $OCF_ROOT/resource.d/heartbeat, as always.

This allows us to run RAs in-place (i.e. directly from the directory
in the source tarballs) without installing them in to the
provider directory.

This is part 1 of the fix for LF bug 2239.