Medium: RA: VirtualDomain: bail out early if config file can't be read during probe (Novell 593988)
If the user placed the libvirt domain config file on shared storage,
then there is no way for us to deduce the correct domain name. Thus we
can't check whether it's running, and the only thing we can assume is
that it's not. Which is ugly, but probably better than throwing an
error.

https://bugzilla.novell.com/show_bug.cgi?id=593988