low: unset CFLAGS in a dash-friendly way
According to the dash man page, the only way to unexport a variable
is to unset it. This also works in bash.

For the record, the reason that CFLAGS is exported from the environment to
configure and configure appends -Werror to the CFLAGS then it will be
re-exported by configure and has been observed to cause trouble when
configure runs a separate configure is the libltdl directory.

See:
Debian Bug #582874 - http://bugs.debian.org/582874
Debian Bug #582875 - http://bugs.debian.org/582875

Signed-off-by: Simon Horman <horms@verge.net.au>