Medium: mysql: if appropriate, connect to master on start
If a new node joins the cluster, a master instance may already be
present. When that is the case, we should start replicating immediately.
Modify start operation so it starts MySQL replication if a master_uname
is available.

--HG--
extra : rebase_source : 05b9e5896ea8725fb25659413eb356eee87a1e2f