Medium: exportfs: add unlock_on_stop parameter
When the resource stops, it's a good idea to relinquish all locks
which the NFS server holds on the exported file system. On Linux,
the NFS server can be forced to do this by echoing the affected path
into the /proc/nfs/nfsd/unlock_filesystem virtual file.

Leave this parameter disabled by default for the following reasons:

- POLS. Don't introduce a new default since people are already
  using this RA.
- The relevant virtual file may not be present on legacy kernels.
- The resource may be running on an OS other than Linux.