ldirectord: low: Receive string must always start with "/"
There are two locations where the receive string can be set.
One corresponding with the real=... configuration directive and one
corresponding with the request=... directive. Unfortunately
the parsing of these was slightly inconsistent which lead to
the truncation of the first letter of the request string
when the real=... syntax is used in conjunction with the simpletcp check.

Reported-by: Niles Ingalls <nilesi@zionsville.lib.in.us>
Signed-off-by: Simon Horman <horms@verge.net.au>