High: SAPInstance: Moved testing of SAP profile directory and START profile to a later stage (only when needed), for more robustness
Very often the SAP profile directory is located on NFS.
Not in any case it is required to access the SAP profile directory.
It is not required during a monitor action, if sapstartsrv is up and responding.
In that situation a stale NFS mount will not cause hanging the monitor action anymore.