High: SAPInstance: Fixed monitor_clone function to ensure enqueue failover, in case of process (not host) failure
RAs in versions <= 2.01 used a Heartbeat 2.0 specific feature to distinquish, if running in master or slave mode.
This is not working with Pacemaker anymore.

Since RA version 2.02 (not in official release) the monitor_clone function is damaged for the case of a local failure of the Standalone Enqueue process.

This patch follows the requirement, that the RA must know be itself, if it is running in master or slave mode.
Also it ensures, that always the slave (Enqueue Replication Server) gets promoted, if the master (Standalone Enqueue Server) fails.